split to function, tape, variable
